This version is fast. It defines macro HANDLE_BOUNDARY so the code can have two version: one which is faster but does not handle boundary of the image; the other is slower but handle boundary of the image.

+++ src/main.cpp
@@ -23,7 +23,7 @@ void main(int argc, char *argv[])
 	int halfWindowSize;
 	int numOfSamples;
 	float SPMAlpha;
-	int gaussianSigma;
+	float gaussianSigma;
 
 	try
 	{
